refactor: :recycle: コントローラー移行（大問）
